chore(refactoring): split functions into individual files, reviewed imports and cleans code
