+++ conventions/custom_convention_handler.py
@@ -1,0 +1,3 @@
+# FUTURE: implement
+def custom_convention():
+    pass
