+++ conventions/no_convention.py
@@ -1,0 +1,6 @@
+from os import system
+from utils import get_text
+
+def just_message():
+    msg = str(input("commit message: "))
+    system("git commit -m '%s'" % msg.capitalize())
